requirements.txt
pygame-ce>=2.5.0
numpy>=1.21.0
torch>=2.0.0
matplotlib>=3.5.0
gymnasium>=1.0.0